Store symbol origin object/dylib indices as u32
Origin::Obj/Dylib carried usize payloads (16-byte enum). Object and
dylib counts fit in u32, so they become u32 (8-byte enum), and the
dynamic-lookup sentinel changes from usize::MAX to u32::MAX
(bind_ordinal now takes u32). Match sites cast the payload back to
usize where they index ctx.objs/ctx.dylibs. Symbol drops from 64 to
56 bytes. Output unchanged; 322 tests pass, large programs link/run.

+++ src/symbol.rs
@@ -9,10 +9,10 @@ pub type SymbolId = usize;
 pub enum Origin {
     /// Not defined by any input file (yet).
     Undef,
-    /// Defined by an object file.
-    Obj(usize),
-    /// Exported by a dylib.
-    Dylib(usize),
+    /// Defined by an object file (index; u32 to keep Symbol small).
+    Obj(u32),
+    /// Exported by a dylib (index).
+    Dylib(u32),
     /// Defined by the linker itself, e.g. `__mh_execute_header`.
     Synthetic,
 }
